// src/scenario/douban/pages/review-detail/types.ts
export interface ReviewDetailData {
  id: string;
  title: string;
  reviewUrl: string;
  authorName: string;
  authorId: string;
  authorUrl: string;
  avatarUrl: string;
  subjectTitle: string;
  subjectUrl: string;
  posterUrl: string;
  rating: number;
  paragraphs: string[];
  date: string;
  location: string;
  readCount: number;
  source: string;
  usefulCount: number;
  uselessCount: number;
  director: string;
  cast: string;
  genre: string;
  region: string;
  releaseDate: string;
}